+++ config/db.js
@@ -1,0 +1,17 @@
+const mongoose = require('mongoose');
+
+require('dotenv').config();
+
+const MONGO_URI = process.env.MONGO_URI || 'mongodb://localhost:27017/MaharaniDB';
+
+const connectDB = async () => {
+  try {
+    const conn = await mongoose.connect(MONGO_URI);
+    console.log(`MongoDB Connected: ${conn.connection.host}`);
+  } catch (error) {
+    console.error(`Error connecting to MongoDB: ${error.message}`);
+    process.exit(1);
+  }
+};
+
+module.exports = connectDB;
